Moved the MoveInfo class from use cases to entities.
